Move HealthReductionCause to the HealthSystem
